+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### [0.3.0] - 2016-03-02
-- Added pt-BR translation
+* Add pt_BR translation
 
 ### [0.2.0] - 2015-10-25
-- Initial public version
+* Initial public version